[ll] core: Reorganize queue related traits
